Added fontawesome
The Icons are under a CC BY 4.0-license, which I satisfy by mentioning them here, and the Fonts are under a SIL OFL 1.1-license, which I satisfy by not renaming the fonts and the files. Because of this it should be okay for me to use it here in my open-source project.