Did some commenting on the code
It was really needed, I should have done this earlier but I was focusing a lot on getting the stuff I needed done then. I'm also trying to generate docs for this but running into some problem - jsdoc no longer has a website and esdoc does not work correctly. I won't commit the docs and code for the docs until it works.