Did more work on classes as plugins
Changed what is exported from modules\plugins\timeclass.js so other code can know it is a class without going through the hassle on checking that, added code for scripts\renderer\fillGrid.js and scripts\renderer\registerplugins.js so that it considers this. Changed scripts\renderer\registerdetails.js so that if a imported module is a function it calls it when adding it to loaded_details, and if not it just adds it.